style : all style change
-기본 틀 다짐
