+++ src/component/Board/filter.css
@@ -1,0 +1,23 @@
+.filter{
+  background: var(--bg-element1);
+  height: 2rem;
+  width: 6rem;
+  border-radius: 4px;
+  display: flex;
+  -webkit-box-align: center;
+  align-items: center;
+  -webkit-box-pack: justify;
+  justify-content: space-between;
+  padding-left: 0.5rem;
+  padding-right: 0.5rem;
+  font-weight: 600;
+  color: var(--text2);
+  font-size: 0.875rem;
+  box-shadow: rgb(0 0 0 / 5%) 0px 0px 4px;
+  cursor: pointer;
+}
+
+.filter svg{
+  width: 1.5rem;
+  height: 1.5rem;
+}
