+++ src/component/Board/recent.css
@@ -1,0 +1,12 @@
+.recent{
+  width: 7rem;
+  display: flex;
+  -webkit-box-align: center;
+  align-items: center;
+  -webkit-box-pack: center;
+  justify-content: center;
+  font-size: 1.125rem;
+  text-decoration: none;
+  color: var(--text3);
+  height: 3rem;
+}
