+++ src/component/Board/tap.css
@@ -1,0 +1,86 @@
+/* [data-component="Tap"]{
+  margin-left: auto;
+  margin-right: auto;
+} */
+
+#tap{
+  margin-top: 1.5rem;
+  display: flex;
+  -webkit-box-pack: justify;
+  justify-content: space-between;
+  -webkit-box-align: center;
+  align-items: center;
+  position: relative;
+}
+
+.selectTap{
+  display: flex;
+  -webkit-box-align: center;
+  align-items: center;
+  position: relative;
+}
+
+.underline.tranding{
+  width: 50%;
+  height: 2px;
+  position: absolute;
+  bottom: 0px;
+  background-color: #343A40;
+  animation: leftUnderLine .5s
+  
+}
+.underline{
+  width: 50%;
+  height: 2px;
+  position: absolute;
+  bottom: 0px;
+  background-color: #343A40;
+
+}
+
+.underline.recent{
+  width: 50%;
+  height: 2px;
+  position: absolute;
+  bottom: 0px;
+  left:50%;
+  background-color: #343A40;
+  animation: rightUnderLine .5s
+}
+
+.select{
+  display: flex;
+  position: relative;
+}
+
+@keyframes rightUnderLine {
+  from{
+    left:0%;
+  }
+  to{
+    left:50%;
+  }
+  
+}
+
+@keyframes leftUnderLine {
+  from{
+    left:50%;
+  }
+  to{
+    left:0%;
+  }
+  
+}
+
+@media (max-width: 1919px){
+  #tap{
+    width: 1376px;
+  }
+}
+
+@media (max-width: 1440px){
+  #tap{
+    width: 1024px;
+  }
+}
