+++ src/component/Board/tranding.ts
@@ -1,0 +1,16 @@
+import Component from "../../core/Component";
+import "./tranding.css";
+export default class Tranding extends Component{
+  setup(){
+  }
+  
+  template(){
+    return `
+    <div class='tranding'>
+      <svg stroke="currentColor" fill="currentColor" stroke-width="0" viewBox="0 0 24 24" height="1em" width="1em" xmlns="http://www.w3.org/2000/svg"><path d="M16 6l2.29 2.29-4.88 4.88-4-4L2 16.59 3.41 18l6-6 4 4 6.3-6.29L22 12V6z"></path></svg>
+      트렌딩
+    </div>`
+  }
+
+}
+
