--- src/component/Tap/tap.css
@@ -1,187 +1,0 @@
-[data-component="Tranding"]{
-  width: 48%;
-  height: 100%;
-  margin: 0px;
-  display: inline-block;
-  color: gray;
-  text-align: center;
-
-}
-
-[data-component="Recent"]{
-  
-  width: 48%;
-  height: 100%;
-  margin: 0px;
-  display: inline-block;
-  color: gray;
-  text-align: center;
-
-}
-
-div.tap{
-  align-items: center;
-}
-
-div.tap > *{
-  display: inline-block;
-  cursor: pointer;
-  height: 40px;
-  font-weight: bold;
-}
-
-[data-component="Submenu"]{
-  float:right;
-  
-}
-.submenu{
-  position: relative;
-  top: 50%;
-}
-
-div.filter{
-  width: 100px;
-  position: relative;
-  top: 30%;
-  transform: translateY(-30%);
-  border: 1px solid whitesmoke;
-  text-align: left;
-  font-size: 17px;
-  font-weight: 600;
-  padding: 6px 15px;
-  transform: scale(0.8);
-  
-}
-div.filter>svg{
-  margin-left: 5px;
-  transform: scale(1.2);
-  float:right;
-}
-
-@keyframes rightUnderLine {
-  from{
-    left:0%;
-  }
-  to{
-    left:50%;
-  }
-}
-
-@keyframes leftUnderLine {
-  from{
-    left:50%;
-  }
-  to{
-    left:0%;
-  }
-}
-
-div.underline{
-  width: 50%;
-  height: 2px;
-  position: relative;
-  bottom: 0px;
-  background: black;
-  left:0%
-}
-
-.selectTap{
-  box-sizing: content-box;
-  width: 10%;
-  padding:10px;
-}
-
- div.underline.left{
-  animation: leftUnderLine 0.5s forwards;
-}
-
-div.underline.right{
-  animation: rightUnderLine 0.5s forwards;
-}
-.selected{
-  color: black;
-}
-
-.recent{
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.tranding{
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-
-}
-
-@keyframes fadein {
-  from{
-    opacity: 0;
-    transform: scale(0.5);
-  }
-  to{
-    opacity: 1;
-    transform: scale(1);
-  }
-  
-}
-
-@keyframes fadeout {
-  from{
-    opacity: 1;
-    transform: scale(1);
-  }
-  to{
-    opacity: 0;
-    transform: scale(0.5);
-    display: none;
-  }
-}
-
-
-[data-component="Dropdown-filter"].fadein{
-  width: 190px;
-  height: 200px;
-
-  animation: fadein .5s forwards;
-  transform-origin: 100% 0px;
-
-  
-}
-
-[data-component="Dropdown-filter"].fadeout{
-  width: 190px;
-  height: 200px;
-  animation: fadeout .5s forwards;
-  transform-origin: 100% 0px;
-
-  
-}
-
-[data-component="Dropdown-submenu"] li:hover{
-
-  color: #12B886;
-  
-}
-
-
-[data-component="Dropdown-submenu"].fadein{
-  width: 190px;
-  height: 200px;
-
-  animation: fadein .5s forwards;
-  transform-origin: 100% 0px;
-
-  
-}
-
-[data-component="Dropdown-submenu"].fadeout{
-  width: 190px;
-  height: 200px;
-  animation: fadeout .5s forwards;
-  transform-origin: 100% 0px;
-
-  
-}
-
