+++ src/component/header/header.css
@@ -1,76 +1,38 @@
+body {
+  background-color: #F8F9FA;
+  box-sizing: border-box;
+}
+
+#root{
+  width: 100%;
+}
+
+[data-component="Header"]{
+  height: 4rem;
+
+}
 #header{
-  position: relative;
-  width:100%;
-  height:4rem;
-  
-}
-
-#header > div{
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-  margin: 0px 8px;
-  cursor:pointer;
-}
-
-div[data-component='Logo']{
-  width:100px;
-  height: 50px;
-  display: inline-block;
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-  
-}
-
-div[data-component='NightMode']{
-  float:right;
-  width: 40px;
-  height: 40px;
-  border-radius: 100%;
-
-}
-div[data-component='NightMode']:hover{
-  background-color: whitesmoke;
-
-}
-div[data-component='NightMode'] > *{
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-  margin-left:auto;
-  margin-right:auto;
-  
-}
-
-
-div[data-component='Search']{
-  float:right;
-  width: 40px;
-  height: 40px;
-  
-  border-radius: 100%;
-
-}
-div[data-component='Search']:hover{
-  background-color: whitesmoke;
-
-}
-
-div[data-component='Search'] > *{
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
   margin-left: auto;
-  margin-right:auto;
+  margin-right: auto;
 }
-div[data-component='Login']{
-  float:right;
-  height: 50px;
+@media (max-width: 1919px){
+  #header{
+    width: 1375px;
+  }
+}
+@media (max-width: 1440px){
+  #header{
+    width: 1024px;
+  }
+}
 
-}
-div[data-component='Login'] > *{
+.header-menu{
+  display: flex;
+  -webkit-box-align: center;
+  align-items: center;
   position: relative;
-  top: 50%;
-  transform: translateY(-50%);
 }
