+++ src/component/header/login.ts
@@ -10,6 +10,4 @@ export default class Login extends Component{
     </div>`
   }
 
-
-  
 }
