+++ src/component/header/search.css
@@ -1,4 +1,16 @@
-.search-btn{
-  width: 17px;
-  height: 17px;
+[data-component="Search"]{
+  display: flex;
+  -webkit-box-align: center;
+  align-items: center;
+  -webkit-box-pack: center;
+  justify-content: center;
+  background: transparent;
+  border: none;
+  width: 2.5rem;
+  height: 2.5rem;
+  outline: none;
+  border-radius: 50%;
+  color: #151921;
+  cursor: pointer;
+  margin-right: 0.5rem;
 }
