+++ src/page/main.ts
@@ -1,13 +1,13 @@
-import Header from "../component/header/header";
+import Board from "component/Board/board";
 import Component from "../core/Component";
 
 class Main extends Component{
   template(){
-    return `<div data-component='Header'></div>`;
+    return `<div data-component='Board'></div>`;
   }
   mounted(){
-    const $header = this.$target.querySelector('[data-component="Header"]');
-    new Header($header, {});
+    const $board = this.$target.querySelector('[data-component="Board"]');
+    new Board($board, {});
   }
 }
